Calling ParserGC

+++ native/core/allocate.c
@@ -49,7 +49,7 @@ void ParserGC_free(void) {
             if (PointerSet_contains(freed_pointers, (size_t)pointer)) {
                 struct VoidPointerListElement *debug_iterator = VoidPointerList_begin(allocated_pointers);
                 while (debug_iterator) {
-                    fprintf(stderr, "%zx\n", pointer);
+                    fprintf(stderr, "%zx\n", (size_t)pointer);
                     debug_iterator = VoidPointerListIterator_next(debug_iterator);
                 }
                 fail_with_message("Double free from ParserGC");
